refactor: move Solidity bindings to solidity-abi and drop go.work
Address review feedback on #1362:

- move the bindings module from `gen/go/solidity-abi` to `solidity-abi`, so
  imports read `github.com/cosmos/ibc/solidity-abi/...` and the module path
  matches its repo-relative directory
- drop the committed `go.work`/`go.work.sum` and gitignore them at the repo
  root, mirroring `link/.gitignore`; restore the `require`/`replace` wiring in
  `e2e/go.mod` for both `link` and `solidity-abi`
- bump `github.com/ethereum/go-ethereum` to v1.17.5 in the bindings module to
  match `e2e` and `link`
- move `generate-contract-bindings.sh` next to the module it writes, at
  `solidity-abi/scripts/`
- rename the `*-gen` Make targets to `*-solidity-abi` and retarget the license
  config and CI path filters, caches, and lint working directories

The generated binding contents are unchanged; regenerating with `make
check-test-apps` reproduces them byte for byte.
