+++ .github/workflows/e2e.yml
@@ -29,15 +29,13 @@ jobs:
           filters: |
             lint:
               - "e2e/**"
-              - "gen/go/solidity-abi/**"
-              - "go.work"
+              - "solidity-abi/**"
               - ".golangci.yml"
               - ".github/workflows/e2e.yml"
             code:
               - "link/**"
               - "e2e/**"
-              - "gen/go/solidity-abi/**"
-              - "go.work"
+              - "solidity-abi/**"
               - "proto/**"
               - "Makefile"
               - ".github/workflows/e2e.yml"
@@ -53,7 +51,7 @@ jobs:
           go-version-file: e2e/go.mod
           cache-dependency-path: |
             e2e/go.sum
-            gen/go/solidity-abi/go.sum
+            solidity-abi/go.sum
       - uses: golangci/golangci-lint-action@ba0d7d2ec06a0ea1cb5fa41b2e4a3ab91d21278a # v9.3.0
         with:
           version: v2.12.2
@@ -61,7 +59,7 @@ jobs:
       - uses: golangci/golangci-lint-action@ba0d7d2ec06a0ea1cb5fa41b2e4a3ab91d21278a # v9.3.0
         with:
           version: v2.12.2
-          working-directory: gen/go/solidity-abi
+          working-directory: solidity-abi
 
   test:
     needs: changes
@@ -75,7 +73,7 @@ jobs:
           cache-dependency-path: |
             link/go.sum
             e2e/go.sum
-            gen/go/solidity-abi/go.sum
+            solidity-abi/go.sum
       - run: make build-link
       - run: docker info
       - run: go test -count=1 ./internal/... ./cmd/...
@@ -114,7 +112,7 @@ jobs:
           cache-dependency-path: |
             link/go.sum
             e2e/go.sum
-            gen/go/solidity-abi/go.sum
+            solidity-abi/go.sum
       - run: docker info
       - run: make test-e2e
 
@@ -127,6 +125,6 @@ jobs:
           go-version-file: e2e/go.mod
           cache-dependency-path: |
             e2e/go.sum
-            gen/go/solidity-abi/go.sum
+            solidity-abi/go.sum
       - run: docker info
       - run: make check-e2e-matrix
