+++ .gitignore
@@ -12,3 +12,6 @@ examples/**/local/
 # besu-to-besu example: generated at setup time
 examples/besu-to-besu/chains/local/
 examples/besu-to-besu/logs/
+
+go.work
+go.work.sum
