+++ e2e/go.mod
@@ -290,6 +290,8 @@ require (
 	github.com/containerd/errdefs/pkg v0.3.0 // indirect
 	github.com/containerd/log v0.1.0 // indirect
 	github.com/containerd/platforms v0.2.1 // indirect
+	github.com/cosmos/ibc/link v0.0.0
+	github.com/cosmos/ibc/solidity-abi v0.0.0
 	github.com/cosmos/solidity-ibc-eureka/packages/go-abigen v0.0.0-20260810020832-a40957eaf878
 	github.com/cpuguy83/dockercfg v0.3.2 // indirect
 	github.com/crate-crypto/go-eth-kzg v1.5.0 // indirect
@@ -343,4 +345,9 @@ require (
 	golang.org/x/sys v0.47.0 // indirect
 	google.golang.org/protobuf v1.36.12-0.20260120151049-f2248ac996af // indirect
 	gopkg.in/yaml.v3 v3.0.1
+)
+
+replace (
+	github.com/cosmos/ibc/link => ../link
+	github.com/cosmos/ibc/solidity-abi => ../solidity-abi
 )
