+++ e2e/internal/e2etest/traffic_evm.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/cosmos/ibc/e2e/internal/harness/chain/evm"
 	"github.com/cosmos/ibc/e2e/internal/harness/environment"
-	"github.com/cosmos/ibc/gen/go/solidity-abi/testerc20"
+	"github.com/cosmos/ibc/solidity-abi/testerc20"
 )
 
 func mustABI(metadata *bind.MetaData) abi.ABI {
