+++ solidity-abi/counter/contract.go
