+++ solidity-abi/iftbatchtransfershim/contract.go
